Refactor WPML Cleanup into WPML module
